+++ src/components/GameOver.tsx
@@ -3,6 +3,7 @@ import './less/game-over.less';
 import React, { ChangeEvent, useState } from 'react';
 import { inject, observer } from 'mobx-react';
 import { Space, Table, Form, Input, Button } from 'antd';
+import { useTranslation } from 'react-i18next';
 
 import InjectionError from './helpers/InjectionError';
 
@@ -36,20 +37,22 @@ const GameOver = (props: GameOverProps) => {
     score
   } = lastResult;
 
+  const { t } = useTranslation('game-over');
+
   const dataSource = [{
-    label: 'Score',
+    label: t('score'),
     value: score,
     key: 'score'
   }, {
-    label: 'Correct count',
+    label: t('correct'),
     value: correctCount,
     key: 'correntCount'
   }, {
-    label: 'Wrong count',
+    label: t('wrong'),
     value: wrongCount,
     key: 'wrongCount'
   }, {
-    label: 'Missed count',
+    label: t('missed'),
     value: missedCount,
     key: 'missedCount'
   }];
@@ -77,24 +80,20 @@ const GameOver = (props: GameOverProps) => {
     gameStore.finish();
   };
 
-  const initialValues = {
-    name
-  };
+  const initialValues = { name };
 
   return (
     <div className="game-over">
       <Space direction="vertical" size="large">
-        <div className="game-over-title">Game over</div>
+        <div className="game-over-title">{t('game-over')}</div>
         <div className="game-over-name">
           <Form name="game-over" layout="inline" initialValues={initialValues} onSubmitCapture={onNameSubmit}>
-            <Form.Item
-              name="name"
-            >
-              <Input onChange={onNameChange} placeholder="Input your name..." />
+            <Form.Item name="name">
+              <Input onChange={onNameChange} placeholder={t('input-your-name')} />
             </Form.Item>
             <Form.Item>
               <Button type="primary" onSubmit={onNameSubmit} onClick={onNameSubmit} disabled={!name.trim()}>
-                Enter
+                {t('enter')}
               </Button>
             </Form.Item>
           </Form>