+++ src/components/Intro.tsx
@@ -2,6 +2,7 @@ import './less/intro.less';
 
 import React from 'react';
 import { Space, Button } from 'antd';
+import { useTranslation } from 'react-i18next';
 
 import AppTitle from './AppTitle';
 
@@ -10,6 +11,7 @@ interface IntroProps extends React.HTMLAttributes<HTMLDivElement> {
 }
 
 const Intro: React.FunctionComponent<IntroProps> = props => {
+  const { t } = useTranslation('intro');
   const { close } = props;
   return (
     <div className="intro">
@@ -19,7 +21,7 @@ const Intro: React.FunctionComponent<IntroProps> = props => {
           <div className="intro-logo" />
           <div className="intro-buttons">
             <Button type="primary" size="large" onClick={close}>
-              Press the button
+              {t('press-the-button')}
             </Button>
           </div>
         </Space>