+++ src/components/Start.tsx
@@ -2,6 +2,7 @@ import './less/start.less';
 
 import React from 'react';
 import { Space, Button } from 'antd';
+import { useTranslation } from 'react-i18next';
 
 import AppTitle from './AppTitle';
 
@@ -14,6 +15,7 @@ interface StartProps extends React.HTMLAttributes<HTMLDivElement> {
 }
 
 const Start: React.FunctionComponent<StartProps> = props => {
+  const { t } = useTranslation('start');
   const { start, canBeStarted } = props;
   return (
     <div className="start">
@@ -21,18 +23,18 @@ const Start: React.FunctionComponent<StartProps> = props => {
         <AppTitle />
         <div className="start-content">
           <div className="start-pane">
-            <h2>Settings</h2>
+            <h2>{t('settings')}</h2>
             <Settings />
           </div>
           <div className="start-pane-divider"></div>
           <div className="start-pane">
-            <h2>Best results</h2>
+            <h2>{t('best-results')}</h2>
             <BestResults />
           </div>
         </div>
         <div className="start-footer">
           <Button type="primary" size="large" danger={true} disabled={!canBeStarted} onClick={start}>
-            Start the game
+            {t('start-the-game')}
           </Button>
         </div>
       </Space>